o  int -> long for a couple of method return types.
git-svn-id: https://svn.apache.org/repos/asf/maven/continuum/branches/id-refactor@499917 13f79535-47bb-0310-9956-ffa450edef68

+++ continuum-api/src/main/java/org/apache/maven/continuum/Continuum.java
@@ -197,7 +197,7 @@ public interface Continuum
      * @return id of the project
      * @throws ContinuumException
      */
-    int addProject( Project project, String executorId )
+    long addProject( Project project, String executorId )
         throws ContinuumException;
 
     /**
@@ -209,7 +209,7 @@ public interface Continuum
      * @return id of the project
      * @throws ContinuumException
      */
-    int addProject( Project project, String executorId, long projectGroupId )
+    long addProject( Project project, String executorId, long projectGroupId )
         throws ContinuumException;
 
     /**
